% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_define_survival_data.R
\name{extract_define_survival_data}
\alias{extract_define_survival_data}
\title{Survival Preparation}
\usage{
extract_define_survival_data(cancer_record_dataset_path, stata_exe_path = NULL)
}
\arguments{
\item{cancer_record_dataset_path}{`[character]` (mandatory, no default)

path to a dataset of cancer records}

\item{stata_exe_path}{`[NULL, character]` (mandatory, no default)

- `NULL`: path is taken from cache, if there; the path is cached when 
  supplied as a `character`
- `character`: full path to Stata executable}
}
\value{
[system] is called internally and that function's output is returned
(an integer).
}
\description{
Run a Stata script on a dataset saved on-disk to prepare a dataset of
on-disk for computing survival.
}
