% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stata_info.R
\name{get_stata_info}
\alias{get_stata_info}
\title{Stata Info}
\usage{
get_stata_info(stata_exe_path)
}
\arguments{
\item{stata_exe_path}{`[NULL, character]` (mandatory, no default)

- `NULL`: path is taken from cache, if there; the path is cached when 
  supplied as a `character`
- `character`: full path to Stata executable}
}
\value{
A `character` vector of the contents of a log file from running a
Stata script in batch mode.
}
\description{
Retrieve Stata version, session, and other information.
}
