% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordcansurvival.R
\name{nordcanstat_survival}
\alias{nordcanstat_survival}
\title{NORDCAN Survival Statistics}
\usage{
nordcanstat_survival(
  cancer_record_dataset,
  national_population_life_table,
  stata_exe_path = NULL
)
}
\arguments{
\item{cancer_record_dataset}{`[data.table]` (mandatory, no default)

the NORDCAN dataset of cancer records after pre-processing}

\item{national_population_life_table}{`[data.table]` (mandatory, no default)

the national population life table}

\item{stata_exe_path}{`[NULL, character]` (mandatory, no default)

- `NULL`: path is taken from cache, if there; the path is cached when 
  supplied as a `character`
- `character`: full path to Stata executable}
}
\value{
Returns a `data.table` of the resulting survival statistics.
}
\description{
Compute survival statistics for NORDCAN. Data is written onto disk and 
Stata is called on the on-disk dataset.
}
