% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_statistics.R
\name{stata_survival_statistics}
\alias{stata_survival_statistics}
\title{Title}
\usage{
stata_survival_statistics(
  stata_exe_path = NULL,
  cancer_record_dataset_path,
  work_dir,
  national_population_life_table_path,
  estimand = "netsurvival"
)
}
\arguments{
\item{estimand}{}
}
\value{

}
\description{
Title
}
\examples{

}
