% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_statistics.R
\name{survival_statistics}
\alias{survival_statistics}
\title{survival_statistics}
\usage{
survival_statistics(stata_exe_path = NULL, cancer_record_dataset_path,
  national_population_life_table_path, estimand = "netsurvival")
}
\arguments{
\item{stata_exe_path}{This is set in the NORDCAN settings by default}

\item{cancer_record_dataset_path}{path to a dataset of cancer records}

\item{national_population_life_table_path}{(mandatory, default by NORDCAN system)}

\item{estimand}{defaults to "netsurvival" the only option so far}
}
\value{

}
\description{
Provides non-parametric estimate of marginal relative survival, which under assumptions can be 
interpreted as marginal net survival. survival_statistics() is calling the Stata stnet program, 
using the Pohar Perme estimator with Brenner weighting to estimate age-standardized net survival.

The beta version provides 5-year survival only.

For some entities estimation is not possible because there are no data in some age stratum used
for weighting. 

In the beta version there is no lower limit to data, execpt there must be at least one observation 
in each strata.
}
\examples{


library(nordcansurvival)
infile <-  paste0(system.file(package = "nordcansurvival"), "/Stata/demo/NCS_NO_anonymous_example_data.dta")
lifetable <- paste0(system.file(package = "nordcansurvival"), "/Stata/demo/NO_2018_lifetable.dta")
nordcancore::set_global_nordcan_settings(
  work_dir = getwd(),
  stat_cancer_record_count_year_first = 1953,
  stat_cancer_record_count_year_last = 2018,
  stat_prevalent_subject_count_year_first = 1953,
  stat_prevalent_subject_count_year_last= 2018,
  stat_survival_follow_up_year_first = 1953,
  stat_survival_follow_up_year_last = 2018
)


survival_statistics(cancer_record_dataset_path  = infile , 
                   national_population_life_table_path = lifetable, 
                   stata_exe_path = "S:/Prog64/STATA/Stata16MP/StataMP-64.exe")


}
