% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_define_survival_data.R
\name{extract_define_survival_data}
\alias{extract_define_survival_data}
\title{extract_define_survival_data}
\usage{
extract_define_survival_data(
  cancer_record_dataset_path,
  stata_exe_path = NULL,
  survival_test_sample,
  survival_trace = FALSE
)
}
\arguments{
\item{cancer_record_dataset_path}{`[character]` (mandatory, no default) path 
to a dataset of cancer records}

\item{stata_exe_path}{This is set in the NORDCAN settings by default}

\item{survival_test_sample}{Logical, if set to TRUE, the function will extract
only a sample of the input cancer record dataset, which contains  
10 percent of Breast, Prostate and colorectal cancers start from 2001.}

\item{survival_trace}{Logical, if set to TRUE, run Stata in debug mode.}
}
\value{
It will generate four dta files for further survival analysis: 
  \itemize{
    \item \emph{survival_file_analysis_survivaltime_05_period_05.dta}
    \item \emph{survival_file_analysis_survivaltime_05_period_10.dta}
    \item \emph{survival_file_analysis_survivaltime_10_period_05.dta}
    \item \emph{survival_file_analysis_survivaltime_10_period_10.dta}
  }
   add additional files: 
  \itemize{
    \item \emph{survival_file_base.dta }
    \item \emph{survival_file_analysis.dta}
   }

Detailed documentation is available at \url{https://github.com/CancerRegistryOf Norway/NORDCAN/wiki/nordcansurvival}
}
\description{
Run a Stata script on a dataset saved on-disk to prepare datasets for analysis 
for computing survival.

survival_file_base.dta  includes all potential records, all relevant variables including strings

survival_file_analysis.dta includes a minimum of variables used for estimation and only observations
for entities were data exists in all strata used for weighting

the dta files can be read in R using the optional haven package
}
