% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stata_info.R
\name{get_stata_info}
\alias{get_stata_info}
\title{get_stata_info}
\usage{
get_stata_info(stata_exe_path)
}
\arguments{
\item{stata_exe_path}{Path to Stata executable. In NORDAN taken from the settings by default.}
}
\value{
Return Stata information to terminal and save as a log file.


}
\description{
Retrieve Stata version, session, and other information for debugging.

This function is for use locally by the user to get informations on Stata installation
and settings , to report in debugging.

The file includes path information and should probably not be distributed outside your local setting.
}
