% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordcansurvival.R
\name{nordcanstat_survival}
\alias{nordcanstat_survival}
\title{NORDCAN Survival Statistics}
\usage{
nordcanstat_survival(
  cancer_record_dataset,
  national_population_life_table,
  stata_exe_path
)
}
\arguments{
\item{cancer_record_dataset}{`[data.table]` (mandatory, no default)

the NORDCAN dataset of cancer records after pre-processing}

\item{national_population_life_table}{`[data.table]` (mandatory, no default)

the national population life table}

\item{stata_exe_path}{`[character]` (mandatory, no default)

full path to Stata executable}
}
\description{
Compute survival statistics for NORDCAN. Data is written onto disk and 
Stata is called on the on-disk dataset.
}
