% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_statistics.R
\name{survival_statistics}
\alias{survival_statistics}
\title{Survival}
\usage{
survival_statistics(
  stata_exe_path = NULL,
  cancer_record_dataset_path,
  national_population_life_table_path,
  estimand = "netsurvival"
)
}
\arguments{
\item{stata_exe_path}{`[NULL, character]` (mandatory, no default)

- `NULL`: path is taken from cache, if there; the path is cached when 
  supplied as a `character`
- `character`: full path to Stata executable}

\item{cancer_record_dataset_path}{`[character]` (mandatory, no default)

path to a dataset of cancer records}

\item{national_population_life_table_path}{`[character]` (mandatory, no default)

path to life table}

\item{estimand}{`[character]` (mandatory, no default)

relative / net survival estimand}
}
\value{
[system] is called internally and that function's output is returned
(an integer).
}
\description{
Run a Stata script on a dataset saved on-disk to produce a table of
survival statistics on-disk.
}
