% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_statistics.R
\name{survival_statistics}
\alias{survival_statistics}
\title{survival_statistics}
\usage{
survival_statistics(
  stata_exe_path = NULL,
  infile,
  lifetable,
  outfile = "survival_statistics",
  estimand = "netsurvival"
)
}
\arguments{
\item{stata_exe_path}{This is set in the NORDCAN settings by default}

\item{infile}{path to a dataset of cancer records}

\item{lifetable}{(mandatory, default by NORDCAN system)}

\item{outfile}{path where to write result file. The outfile must be a set of 
    c("survival_statistics_standardised_survivaltime_05_period_05",
    "survival_statistics_standardised_survivaltime_05_period_10",
    "survival_statistics_standardised_survivaltime_10_period_05",
    "survival_statistics_standardised_survivaltime_10_period_10",
    "survival_statistics_agespecific_survivaltime_05_period_05",
    "survival_statistics_agespecific_survivaltime_05_period_10",
    "survival_statistics_agespecific_survivaltime_10_period_05",
    "survival_statistics_agespecific_survivaltime_10_period_10"), 

    and it was set up by the "output_objects" of "nordcanepistats::nordcan_statistics_tables"}

\item{estimand}{defaults to "netsurvival" the only option so far}
}
\value{
survival analysis output in .csv (UTF-8) and .dta format.
}
\description{
Provides non-parametric estimate of marginal relative survival, which under assumptions can be 
interpreted as marginal net survival. survival_statistics() is calling the Stata stnet program, 
using the Pohar Perme estimator with Brenner weighting to estimate age-standardized net survival.

Detailed documentation is available at https://github.com/CancerRegistryOf Norway/NORDCAN/wiki/nordcansurvival
}
